requirements.txt
anthropic>=0.40.0
requests>=2.31.0
pyyaml>=6.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
pytrends>=4.9.2
jinja2>=3.1.2
sendgrid>=6.11.0
python-dotenv>=1.0.0
lxml>=5.0.0
